Cross compiling works (not GNU standard!)
It's ugly, but it works for me.

$ USTL_CROSS_COMPILE=zynq ./configure

Then the usual "make && make check"

Copy the built stuff to the target board. The Makefile tells you the
test command to run once you're there.

"make install" is broken, but IMHO that's a good thing since it's not
for this machine anyway.